webapps/app-st/pyproject.toml
[project]
name = "app-st"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "streamlit>=1.45.1",
]
